refactor (OutputView) : 좌표계 생성 부분을 다른 Class로 분리함
- OutputView의 대부분이 좌표계 생성을 이루는 코드로 이루고 있어 다른 클래스로 분리함
